v1/galion/services/document-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
kafka-python==2.0.2
redis==5.0.1
python-magic==0.4.27
aiofiles==23.2.1
prometheus-client==0.19.0

